Switch topics to Stream. Remove Feed and FeedList
