--- src/components/FeedList/FeedList.scss
@@ -1,80 +1,0 @@
-.FeedList-container {
-  position: relative;
-  padding: 15px 0;
-}
-
-.FeedListItems {
-  margin: 0 auto;
-  overflow: auto; // prevent margin collapse
-}
-
-.FeedListItem {
-  margin: 0 $space-4x $space-6x;
-  box-shadow: 0px 4px 10px rgb(42 64 89 / 30%);
-}
-
-.expanded {
-  margin: 0 $space-2x $space-6x;
-}
-
-.sticky-show {
-  opacity: 0;
-}
-
-.tabbar-normal {
-  margin-bottom: 15px;
-}
-
-@keyframes append-animate {
-  from {
-    opacity: 0;
-  }
-  to {
-    opacity: 1;
-  }
-}
-
-.tabbar-sticky {
-  position: fixed;
-  top: 77px;
-  z-index: 22;
-  background-color: rgba(255, 255, 255, 1);
-  border-radius: 5px;
-  box-shadow: 0px 4px 10px rgb(42 64 89 / 30%);
-  animation: append-animate .3s ease;
-}
-
-
-/* animate new box */
-.new-box {
-  animation: append-animate .3s linear;
-}
-
-.collapsedState {
-  border: 1px solid red;
-}
-
-@media screen and (max-width: 600px) {
-  .FeedListItem {
-    margin: 0 10px 10px 10px;
-  }
-
-  .tabbar-sticky {
-    z-index: 22;
-  }
-}
-
-@media screen and (max-width: 425px) {
-  .FeedListItem {
-    padding-top: 10px;
-    margin: 0 0 10px 5px;
-  }
-}
-
-.stream-grid {
-  display: grid;
-  grid-template-columns: 1fr 1fr;
-  justify-items: center;
-  align-items: center;
-  gap: 8px;
-}
