--- src/components/FeedList/TabBar/TabBar.scss
@@ -1,102 +1,0 @@
-.bar {
-  width: 100%;
-  padding: 0 $space-4x;
-  display: flex;
-  align-content: stretch;
-  height: $space-7x;
-  margin: 3px 0;
-}
-
-.tabs {
-  overflow: hidden;
-  flex-grow: 1;
-  > span {
-    cursor: pointer;
-  }
-}
-
-.tab {
-  composes: caption-lt-lg from 'css/typography.scss';
-  padding: 0 15px;
-  height: 28px;
-  line-height: 28px;
-  display: inline-block;
-}
-
-.tab-active {
-  composes: tab;
-  // Extract typography? 1px from .HDR-MINOR
-  @include font-bold;
-  color: $color-rhino;
-  font-size: 15px;
-  background-color: rgba(0, 118, 223, .1);
-  border-radius: 4px;
-  color: rgba(0, 118, 223, 1.0);
-}
-
-.sorter {
-  font-size: 14px;
-}
-
-.sorter-label {
-  line-height: 28px;
-  white-space: nowrap;
-  composes: caption-lt-lg from 'css/typography.scss';
-
-  > span {
-    vertical-align: middle;
-    margin-left: $space-1x;
-  }
-}
-
-.filterLabel {
-  display: none;
-}
-
-@media screen and (max-width: 600px) {
-  .tab {
-    display: block;
-  }
-
-  .filterLabel {
-    display: inline-block;
-  }
-
-  .bar {
-    justify-content: space-between;
-    position: relative;
-    z-index: 10;
-  }
-
-  .filterLabel {
-    font-size: 14px;
-    height: 30px;
-    line-height: 30px;
-
-    span {
-      position: relative;
-      top: 7px;
-    }
-  }
-
-  .tabs {
-    max-width: 200px;
-    height: 30px;
-    overflow: hidden;
-    border-radius: 5px;
-    padding: 0 5px 10px 5px;
-    top: -3px;
-    position: relative;
-    color: rgba(125, 140, 157, 1.000);
-
-    strong {
-      text-transform: capitalize;
-    }
-
-    &:hover {
-      height: 150px;
-      box-shadow: 0px 4px 56px 0px $shadow-color;
-      background-color: #fff;
-    }
-  }
-}
