--- src/components/FeedList/TabBar/index.js
@@ -1,3 +1,0 @@
-import component from './TabBar'
-
-export default component
