--- src/routes/Feed/Feed.scss
@@ -1,21 +1,0 @@
-.create-group-prompt {
-  text-align: center;
-  font-size: 30px;
-  align-items: center;
-  justify-content: center;
-  color: $color-rhino-50;
-  padding-left: 100px;
-  padding-right: 100px;
-}
-
-.sidebar-image {
-  height: 150px;
-  width: 100%;
-  background: no-repeat center;
-  background-size: contain;
-}
-
-.button {
-  font-size: 14px;
-  margin-bottom: 30px;
-}
