--- src/routes/Feed/index.js
@@ -1,4 +1,0 @@
-import component from './Feed'
-import connector from './Feed.connector'
-
-export default connector(component)
