+++ src/routes/GroupExplorer/GroupSearch/GroupSearch.scss
@@ -105,7 +105,6 @@
   overflow: visible;
 }
 
-// :DUPE: @see FeedList.scss
 .card-item {
   margin: 0 $space-4x $space-6x;
 }
