+++ src/routes/Stream/Stream.connector.js
@@ -1,27 +1,33 @@
 import { push } from 'connected-react-router'
 import { connect } from 'react-redux'
 import { get, isEmpty } from 'lodash/fp'
-import { FETCH_POSTS } from 'store/constants'
+import { FETCH_POSTS, FETCH_TOPIC, FETCH_GROUP_TOPIC } from 'store/constants'
 import getMe from 'store/selectors/getMe'
 import getMyMemberships from 'store/selectors/getMyMemberships'
+import getTopicForCurrentRoute from 'store/selectors/getTopicForCurrentRoute'
+import {
+  fetchGroupTopic,
+  fetchPosts,
+  fetchTopic,
+  getCustomView,
+  getHasMorePosts,
+  getPosts
+} from 'routes/Stream/Stream.store'
 import getRouteParam from 'store/selectors/getRouteParam'
+import getGroupTopicForCurrentRoute from 'store/selectors/getGroupTopicForCurrentRoute'
 import getQuerystringParam from 'store/selectors/getQuerystringParam'
 import changeQuerystringParam from 'store/actions/changeQuerystringParam'
 import getGroupForCurrentRoute from 'store/selectors/getGroupForCurrentRoute'
-import {
-  fetchPosts,
-  getCustomView,
-  getPosts,
-  getHasMorePosts
-} from 'components/FeedList/FeedList.store'
 import { updateUserSettings } from 'routes/UserSettings/UserSettings.store'
 import presentPost from 'store/presenters/presentPost'
 import respondToEvent from 'store/actions/respondToEvent'
+import isPendingFor from 'store/selectors/isPendingFor'
+import toggleGroupTopicSubscribe from 'store/actions/toggleGroupTopicSubscribe'
 
 import { createPostUrl } from 'util/navigation'
 
 export function mapStateToProps (state, props) {
-  let group
+  let group, topic, groupTopic
   let groupId = 0
 
   const groupSlug = getRouteParam('groupSlug', state, props)
@@ -41,6 +47,16 @@ export function mapStateToProps (state, props) {
   const customViewTopics = customView?.topics
   const viewName = customViewName
   const viewIcon = customViewIcon
+  const topicName = getRouteParam('topicName', state, props)
+  const topicLoading = isPendingFor([FETCH_TOPIC, FETCH_GROUP_TOPIC], state)
+
+  if (groupSlug) {
+    groupTopic = getGroupTopicForCurrentRoute(state, props)
+    groupTopic = groupTopic && { ...groupTopic.ref, group: groupTopic.group, topic: groupTopic.topic }
+  }
+  if (topicName) {
+    topic = getTopicForCurrentRoute(state, props)
+  }
 
   const context = getRouteParam('context', state, props)
   const view = getRouteParam('view', state, props)
@@ -61,10 +77,12 @@ export function mapStateToProps (state, props) {
   const fetchPostsParam = {
     activePostsOnly,
     context,
+    topicName,
     filter: postTypeFilter,
     slug: groupSlug,
     search,
     sortBy,
+    topic: topic?.id,
     topics: customViewTopics?.toModelArray().map(t => t.id) || [],
     types: customPostTypes
   }
@@ -87,10 +105,15 @@ export function mapStateToProps (state, props) {
     postTypeFilter,
     posts,
     querystringParams,
+    postsTotal: get('postsTotal', groupSlug ? groupTopic : topic),
+    followersTotal: get('followersTotal', groupSlug ? groupTopic : topic),
     routeParams,
     search,
     selectedPostId: getRouteParam('postId', state, props),
     sortBy,
+    topicLoading,
+    topicName,
+    topic,
     view,
     viewIcon,
     viewName,
@@ -99,6 +122,8 @@ export function mapStateToProps (state, props) {
 }
 
 export function mapDispatchToProps (dispatch, props) {
+  const groupSlug = getRouteParam('groupSlug', null, props)
+  const topicName = getRouteParam('topicName', null, props)
   const updateSettings = (params) => dispatch(updateUserSettings(params))
   const routeParams = get('match.params', props)
   const querystringParams = getQuerystringParam(['s', 't'], null, props)
@@ -124,6 +149,15 @@ export function mapDispatchToProps (dispatch, props) {
     fetchPosts: param => offset => {
       return dispatch(fetchPosts({ offset, ...param }))
     },
+    fetchTopic: () => {
+      if (groupSlug && topicName) {
+        return dispatch(fetchGroupTopic(topicName, groupSlug))
+      } else if (topicName) {
+        return dispatch(fetchTopic(topicName))
+      }
+    },
+    toggleGroupTopicSubscribe: groupTopic =>
+      dispatch(toggleGroupTopicSubscribe(groupTopic)),
     newPost: () => dispatch(push(createPostUrl(routeParams, querystringParams)))
   }
 }
