+++ src/routes/Stream/Stream.scss
@@ -26,7 +26,6 @@
   border: 1px solid red;
 }
 
-// :DUPE: @see FeedList.scss
 .card-item {
   margin: 0 $space-4x $space-6x;
   box-shadow: 0 4px 10px rgba(35, 65, 91, .2);
